+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_pos/AnnotatePositionCommandLineParser.java
@@ -24,12 +24,18 @@ public class AnnotatePositionCommandLineParser extends JannovarAnnotationCommand
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.ANNOTATE_POSITION;
 
 		if (cmd.hasOption("help")) {
 			printHelp();
 			throw new HelpRequestedException();
 		}
+
+		if (cmd.hasOption("verbose"))
+			result.verbosity = 2;
+		if (cmd.hasOption("very-verbose"))
+			result.verbosity = 3;
 
 		String args[] = cmd.getArgs(); // get remaining arguments
 		if (args.length < 3)