+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotateVCFCommandLineParser.java
@@ -23,12 +23,18 @@ public class AnnotateVCFCommandLineParser extends JannovarAnnotationCommandLineP
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.ANNOTATE_VCF;
 
 		if (cmd.hasOption("help")) {
 			printHelp();
 			throw new HelpRequestedException();
 		}
+
+		if (cmd.hasOption("verbose"))
+			result.verbosity = 2;
+		if (cmd.hasOption("very-verbose"))
+			result.verbosity = 3;
 
 		result.jannovarFormat = cmd.hasOption("janno");