+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/download/DownloadCommand.java
@@ -29,14 +29,14 @@ public final class DownloadCommand extends JannovarCommand {
 	@Override
 	public void run() throws JannovarException {
 		System.err.println("Options");
-		options.print();
+		options.print(System.err);
 
 		DataSourceFactory factory = new DataSourceFactory(options, options.dataSourceFiles);
 		for (String name : options.dataSourceNames) {
 			System.err.println("Downloading/parsing for data source \"" + name + "\"");
-			JannovarData data = factory.getDataSource(name).getDataFactory().build(options.downloadPath);
+			JannovarData data = factory.getDataSource(name).getDataFactory()
+					.build(options.downloadPath, options.printProgressBars);
 			String filename = PathUtil.join(options.downloadPath, name.replace('/', '_').replace('\\', '_') + ".ser");
-			System.err.println("Serializing to " + filename);
 			JannovarDataSerializer serializer = new JannovarDataSerializer(filename);
 			serializer.save(data);
 		}