+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/download/DownloadCommandLineParser.java
@@ -45,6 +45,9 @@ public final class DownloadCommandLineParser {
 	private void initializeParser() {
 		options = new Options();
 		options.addOption(OptionBuilder.withDescription("show this help").withLongOpt("help").create("h"));
+		options.addOption(OptionBuilder.withDescription("create verbose output").withLongOpt("verbose").create("v"));
+		options.addOption(OptionBuilder.withDescription("create very verbose output").withLongOpt("very-verbose")
+				.create("vv"));
 		options.addOption(OptionBuilder.withDescription("INI file with data source list").hasArgs(1)
 				.withLongOpt("data-source-list").create("s"));
 		options.addOption(OptionBuilder
@@ -82,12 +85,18 @@ public final class DownloadCommandLineParser {
 
 		// Fill the resulting JannovarOptions.
 		JannovarOptions result = new JannovarOptions();
+		result.printProgressBars = true;
 		result.command = JannovarOptions.Command.DOWNLOAD;
 
 		if (cmd.hasOption("help")) {
 			printHelp();
 			throw new HelpRequestedException();
 		}
+
+		if (cmd.hasOption("verbose"))
+			result.verbosity = 2;
+		if (cmd.hasOption("very-verbose"))
+			result.verbosity = 3;
 
 		if (cmd.hasOption("data-dir"))
 			result.downloadPath = cmd.getOptionValue("data-dir");